+++ Package.swift
@@ -6,6 +6,7 @@ let package = Package(
   name: "FKKit",
   platforms: [
     .iOS(.v15),
+    .macOS(.v10_15),
   ],
   products: [
     .library(name: "FKUIKit", targets: ["FKUIKit"]),